Clean up index items collection
