+++ .eleventy.js
@@ -4,9 +4,10 @@ const pluginNavigation = require('@11ty/eleventy-navigation');
 const markdownIt = require('markdown-it');
 const markdownItAnchor = require('markdown-it-anchor');
 const {
-  // getSlugMap,
+  getSlugMap,
   getSortedCollection,
   // getTagList
+  indexItems,
 } = require('./src/_11ty/collections');
 const {
   htmlDateString,
@@ -50,46 +51,8 @@ module.exports = function (eleventyConfig) {
   eleventyConfig.addCollection('sortedEntries', collectionApi =>
     getSortedCollection(collectionApi, 'entries'),
   );
-  eleventyConfig.addCollection('indexItems', collectionApi => {
-    const allItems = collectionApi.getFilteredByGlob([
-      'src/entries/*.md',
-      'src/collections/*.md',
-      'src/series/*.md',
-    ]);
-
-    // Get arrays of the entries that belong to collections and series
-    // (use a Set to prevent duplicate values)
-    const entriesThatBelongToCollections = [
-      ...new Set(
-        collectionApi
-          .getFilteredByGlob('src/collections/*.md')
-          .map(collection => collection.data.entries)
-          .flat(),
-      ),
-    ];
-    const entriesThatBelongToSeries = [
-      ...new Set(
-        collectionApi
-          .getFilteredByGlob('src/series/*.md')
-          .map(collection => collection.data.entries)
-          .flat(),
-      ),
-    ];
-
-    // Filter out entries that belong to a collection or series
-    const indexItems = allItems
-      .filter(item => !entriesThatBelongToCollections.includes(item.data.title))
-      .filter(item => !entriesThatBelongToSeries.includes(item.data.title));
-
-    // Sort items alphabetically by title
-    indexItems.sort((a, b) => {
-      const aTitle = a.data.title;
-      const bTitle = b.data.title;
-      return aTitle < bTitle ? -1 : aTitle > bTitle ? 1 : 0;
-    });
-
-    return indexItems;
-  });
+  eleventyConfig.addCollection('entriesMap', collectionApi => getSlugMap(collectionApi, 'entries'));
+  eleventyConfig.addCollection('indexItems', indexItems);
 
   /**
    * Add custom filters
