+++ src/_11ty/collections/index.js
@@ -1,9 +1,11 @@
 const getSlugMap = require('./getSlugMap');
 const getSortedCollection = require('./getSortedCollection');
 const getTagList = require('./getTagList');
+const indexItems = require('./indexItems');
 
 module.exports = {
   getSlugMap,
   getSortedCollection,
   getTagList,
+  indexItems,
 };
