+++ src/_11ty/collections/indexItems.js
@@ -1,0 +1,40 @@
+module.exports = collectionApi => {
+  const allItems = collectionApi.getFilteredByGlob([
+    'src/entries/*.md',
+    'src/collections/*.md',
+    'src/series/*.md',
+  ]);
+
+  // Get arrays of the entries that belong to collections and series
+  // (use a Set to prevent duplicate values)
+  const entriesThatBelongToCollections = [
+    ...new Set(
+      collectionApi
+        .getFilteredByGlob('src/collections/*.md')
+        .map(collection => collection.data.entries)
+        .flat(),
+    ),
+  ];
+  const entriesThatBelongToSeries = [
+    ...new Set(
+      collectionApi
+        .getFilteredByGlob('src/series/*.md')
+        .map(collection => collection.data.entries)
+        .flat(),
+    ),
+  ];
+
+  // Filter out entries that belong to a collection or series
+  const indexItems = allItems
+    .filter(item => !entriesThatBelongToCollections.includes(item.fileSlug))
+    .filter(item => !entriesThatBelongToSeries.includes(item.fileSlug));
+
+  // Sort items alphabetically by title
+  indexItems.sort((a, b) => {
+    const aTitle = a.data.title;
+    const bTitle = b.data.title;
+    return aTitle < bTitle ? -1 : aTitle > bTitle ? 1 : 0;
+  });
+
+  return indexItems;
+};
