+++ src/collections/css.md
@@ -1,13 +1,13 @@
 ---
 layout: layouts/collection/index.njk
 title: CSS
-date: 2020-10-16
+date: 2020-10-23
 link:
 tags:
 entries:
-  - The Box Model
-  - Cascade, Inheritance, & Specificity
-  - CSS Preprocessing
+  - the-box-model
+  - cascade-inheritance-and-specificity
+  - css-preprocessing
 ---
 
 [description]
